basic routing done, no server needed
