+++ package.json
@@ -13,6 +13,7 @@
     "nodemon": "^2.0.7",
     "react": "^17.0.2",
     "react-dom": "^17.0.2",
+    "react-router-dom": "^5.2.0",
     "react-scripts": "4.0.3",
     "semantic-ui-css": "2.4.1",
     "semantic-ui-react": "2.0.3",
