--- server/data.json
@@ -1,146 +1,0 @@
-[
-    {
-        "name": "МФТИ",
-        "dir": "Прикладные математика и физика",
-        "price": 270000,
-        "years": 4,
-        "points": 258,
-        "subjects": "физика",
-        "id": 0
-    },
-    {
-        "name": "МИРЭА",
-        "dir": "Информатика и вычислительная техника",
-        "price": 177000,
-        "years": 4,
-        "points": 230,
-        "subjects": "физика",
-        "id": 1
-    },
-    {
-        "name": "МИРЭА",
-        "dir": "Прикладная информатика",
-        "price": 177000,
-        "years": 4,
-        "points": 260,
-        "subjects": "информатика",
-        "id": 2
-    },
-    {
-        "name": "МИРЭА",
-        "dir": "Информационные системы и технологии",
-        "price": 177000,
-        "years": 4,
-        "points": 250,
-        "subjects": "физика",
-        "id": 3
-    },
-    {
-        "name": "МТУСИ",
-        "dir": "Информатика и вычислительная техника",
-        "price": 80000,
-        "years": 4,
-        "points": 228,
-        "subjects": "физика",
-        "id": 4
-    },
-    {
-        "name": "ВШЭ",
-        "dir": "Информатика и вычислительная техника",
-        "price": 390000,
-        "years": 4,
-        "points": 275,
-        "subjects": "физика",
-        "id": 5
-    },
-    {
-        "name": "МГТУ",
-        "dir": "Информатика и вычислительная техника",
-        "price": 285000,
-        "years": 4,
-        "points": 255,
-        "subjects": "информатика",
-        "id": 6
-    },
-    {
-        "name": "МГТУ",
-        "dir": "Прикладная информатика",
-        "price": 285000,
-        "years": 4,
-        "points": 243,
-        "subjects": "информатика",
-        "id": 7
-    },
-    {
-        "name": "МИСиС",
-        "dir": "Информатика и вычислительная техника",
-        "price": 228000,
-        "years": 4,
-        "points": 270,
-        "subjects": "информатика",
-        "id": 8
-    },
-    {
-        "name": "Плеханов",
-        "dir": "Прикладная информатика",
-        "price": 120000,
-        "years": 4,
-        "points": 210,
-        "subjects": "информатика",
-        "id": 9
-    },
-    {
-        "name": "Сеченовский",
-        "dir": "Информационные системы и технологии",
-        "price": 242000,
-        "years": 4,
-        "points": 174,
-        "subjects": "физика",
-        "id": 10
-    },
-    {
-        "name": "МЭИ",
-        "dir": "Информатика и вычислительная техника",
-        "price": 81000,
-        "years": 5,
-        "points": 130,
-        "subjects": "информатика",
-        "id": 11
-    },
-    {
-        "name": "МАИ",
-        "dir": "Информатика и вычислительная техника",
-        "price": 65000,
-        "years": 4,
-        "points": 180,
-        "subjects": "физика",
-        "id": 12
-    },
-    {
-        "name": "МАИ",
-        "dir": "Информационные системы и технологии",
-        "price": 80000,
-        "years": 4,
-        "points": 183,
-        "subjects": "физика",
-        "id": 13
-    },
-    {
-        "name": "МИЭТ",
-        "dir": "Информатика и вычислительная техника",
-        "price": 145000,
-        "years": 4,
-        "points": 228,
-        "subjects": "физика",
-        "id": 14
-    },
-    {
-        "name": "МИФИ",
-        "dir": "Информатика и вычислительная техника",
-        "price": 275000,
-        "years": 4,
-        "points": 285,
-        "subjects": "физика",
-        "id": 15
-    }
-]
